Scrapper update
NuGet included for scrapper to work.

Still need to work on title retrieval process and enter it pass it onto the scrapper. Scrapper still needs to collect the information and then pass it on to the createBoxes script to further implement the information and display it on the gui that is levitating.